# apidocs/swagger.json
{
  "consumes": [
    "application/json",
    "text/plain"
  ],
  "produces": [
    "application/json",
    "text/plain"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes the Supergiant Control API.",
    "title": "Supergiant Control API.",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "2.1.2"
  },
  "basePath": "/",
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "List all cloud accounts.",
        "operationId": "listAccounts",
        "responses": {
          "200": {
            "$ref": "#/responses/listAccountsResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create a new cloud account.",
        "operationId": "createAccount",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CloudAccount"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/accounts/{accountName}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get a cloud account.",
        "operationId": "getAccount",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AccountName",
            "name": "accountName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "accounts"
        ],
        "summary": "Update a cloud account.",
        "operationId": "updateAccount",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AccountName",
            "name": "accountName",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CloudAccount"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Delete a cloud account.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AccountName",
            "name": "accountName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/accounts/{accountName}/regions": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get region details a given cloud account.",
        "operationId": "getAccountRegions",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AccountName",
            "name": "accountName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountRegionsResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/accounts/{accountName}/regions/{region}/az": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get availability zone details a given cloud account.",
        "operationId": "getAccountZones",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AccountName",
            "name": "accountName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Region",
            "name": "region",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountZonesResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/accounts/{accountName}/regions/{region}/az/{az}/types": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get virtual machine details for a given cloud account.",
        "operationId": "getAccountTypes",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AccountName",
            "name": "accountName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Region",
            "name": "region",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "AZ",
            "name": "az",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountTypesResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
    "/v1/api/kubeprofiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List profiles.",
        "operationId": "listProfiles",
        "responses": {
          "200": {
            "$ref": "#/responses/listProfilesResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create a profile.",
        "operationId": "createProfile",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/kubeprofiles/{kubeprofileID}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get a profile.",
        "operationId": "getProfile",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "KubeprofileID",
            "name": "kubeprofileID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/profileResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Addresses": {
      "type": "object",
      "title": "Addresses uses cidr to define an ip list.",
      "properties": {
        "cidr": {
          "type": "string",
          "x-go-name": "CIDR"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "BasicAuthUser": {
      "description": "https://kubernetes.io/docs/reference/access-authn-authz/authentication/#static-password-file",
      "type": "object",
      "title": "BasicAuthUser represents an entry of the static password file.",
      "properties": {
        "groups": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Groups"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "CloudAccount": {
      "description": "CloudAccount is settings of account in public or private cloud (e.g. AWS, vCenter)\nName should be unique.",
      "type": "object",
      "properties": {
        "credentials": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Credentials"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "provider": {
          "$ref": "#/definitions/Name"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/model"
    },
    "CloudSpecificSettings": {
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "ErrorCode": {
      "type": "integer",
      "format": "int64",
      "x-go-package": "github.com/supergiant/control/pkg/sgerrors"
    },
...
    "Message": {
      "description": "Message is used in response body to display separate messages to user and developers",
      "type": "object",
      "properties": {
        "devMessage": {
          "description": "DevMessage should be used to display message to developer on API/console",
          "type": "string",
          "x-go-name": "DevMessage"
        },
        "errorCode": {
          "$ref": "#/definitions/ErrorCode"
        },
        "moreInfo": {
          "description": "MoreInfo should be a link to supergiant documentation to display common problems",
          "type": "string",
          "x-go-name": "MoreInfo"
        },
        "userMessage": {
          "description": "UserMessage should be used to display message to user on UI/CLI",
          "type": "string",
          "x-go-name": "UserMessage"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/message"
    },
...
    "Name": {
      "type": "string",
      "x-go-package": "github.com/supergiant/control/pkg/clouds"
    },
...
    "NodeProfile": {
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "Profile": {
      "type": "object",
      "properties": {
        "K8SVersion": {
          "type": "string"
        },
        "addons": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Addons"
        },
        "arch": {
          "type": "string",
          "x-go-name": "Arch"
        },
        "cidr": {
          "type": "string",
          "x-go-name": "CIDR"
        },
        "cloudSpecificSettings": {
          "$ref": "#/definitions/CloudSpecificSettings"
        },
        "dockerVersion": {
          "type": "string",
          "x-go-name": "DockerVersion"
        },
        "exposedAddresses": {
          "description": "ExposedAddresses is a list of cidr/port pairs that will be exposes\nby cloud provider security groups.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Addresses"
          },
          "x-go-name": "ExposedAddresses"
        },
        "flannelVersion": {
          "type": "string",
          "x-go-name": "FlannelVersion"
        },
        "helmVersion": {
          "type": "string",
          "x-go-name": "HelmVersion"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "k8sApiPort": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "K8SAPIPort"
        },
        "k8sServicesCIDR": {
          "type": "string",
          "x-go-name": "K8SServicesCIDR"
        },
        "masterProfiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NodeProfile"
          },
          "x-go-name": "MasterProfiles"
        },
        "networkProvider": {
          "type": "string",
          "x-go-name": "NetworkProvider"
        },
        "networkType": {
          "type": "string",
          "x-go-name": "NetworkType"
        },
        "nodesProfiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NodeProfile"
          },
          "x-go-name": "NodesProfiles"
        },
        "operatingSystem": {
          "type": "string",
          "x-go-name": "OperatingSystem"
        },
        "password": {
          "description": "DEPRECATED: should be a part of the static auth",
          "type": "string",
          "x-go-name": "Password"
        },
        "provider": {
          "$ref": "#/definitions/Name"
        },
        "publicKey": {
          "type": "string",
          "x-go-name": "PublicKey"
        },
        "rbacEnabled": {
          "type": "boolean",
          "x-go-name": "RBACEnabled"
        },
        "region": {
          "type": "string",
          "x-go-name": "Region"
        },
        "staticAuth": {
          "$ref": "#/definitions/StaticAuth"
        },
        "subnets": {
          "description": "This field is AWS specific, mapping AZ -\u003e subnet",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Subnets"
        },
        "ubuntuVersion": {
          "type": "string",
          "x-go-name": "UbuntuVersion"
        },
        "user": {
          "description": "TODO: get rid of its usage\nDEPRECATED: should be a part of the static auth",
          "type": "string",
          "x-go-name": "User"
        },
        "zone": {
          "type": "string",
          "x-go-name": "Zone"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "Region": {
      "description": "Region represents",
      "type": "object",
      "properties": {
        "AvailableSizes": {
          "description": "API specific IDs for a node sizes/type",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "description": "API specific ID, e.g. t2.micro",
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "description": "Human readable name, e.g. New York City 1 or EU West 1 Frankfurt",
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/account"
    },
    "RegionSizes": {
      "description": "RegionSizes represents aggregated information about available regions/azs and node sizes/types",
      "type": "object",
      "properties": {
        "provider": {
          "$ref": "#/definitions/Name"
        },
        "regions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Region"
          },
          "x-go-name": "Regions"
        },
        "sizes": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          },
          "x-go-name": "Sizes"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/account"
    },
...
    "StaticAuth": {
      "type": "object",
      "title": "StaticAuth represents tokens and basic authentication credentials.",
      "properties": {
        "basicAuth": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BasicAuthUser"
          },
          "x-go-name": "BasicAuth"
        },
        "tokens": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TokenAuthUser"
          },
          "x-go-name": "Tokens"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
    "TokenAuthUser": {
      "description": "https://kubernetes.io/docs/reference/access-authn-authz/authentication/#static-token-file",
      "type": "object",
      "title": "BasicAuthUser represents an entry of the static token file.",
      "properties": {
        "groups": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Groups"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "token": {
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/profile"
    },
...
  },
  "responses": {
    "accountRegionsResponse": {
      "description": "accountRegionsResponse returns a list of supported regions for a given account.",
      "schema": {
        "$ref": "#/definitions/RegionSizes"
      }
    },
    "accountResponse": {
      "description": "accountResponse contains representations of a cloud accounts.",
      "schema": {
        "$ref": "#/definitions/CloudAccount"
      }
    },
    "accountTypesResponse": {
      "description": "accountTypesResponse contains a list of supported machine types for a given account.",
      "schema": {
        "type": "array",
        "items": {
          "type": "string"
        }
      }
    },
    "accountZonesResponse": {
      "description": "accountZonesResponse contains a list of supported availability zones for a given account.",
      "schema": {
        "type": "array",
        "items": {
          "type": "string"
        }
      }
    },
...
    "emptyResponse": {
      "description": "emptyResponse represents an empty http response."
    },
    "errorResponse": {
      "description": "errorResponse represents an error response.",
      "schema": {
        "$ref": "#/definitions/Message"
      }
    },
...
    "listAccountsResponse": {
      "description": "listAccountsResponse contains a list of cloud accounts.",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/CloudAccount"
        }
      }
    },
...
    "listProfilesResponse": {
      "description": "listProfilesResponse returns all kube profiles.",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Profile"
        }
      }
    },
...
    "profileResponse": {
      "description": "profileResponse contains representations of a kube profile.",
      "schema": {
        "$ref": "#/definitions/Profile"
      }
    },
...
  },
  "securityDefinitions": {
    "AuthorizationHeaderToken": {
      "description": "This scheme uses the header \"Authorization: Bearer {JWT token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}